FIX: Minor variable casting warnings fixed. FIX: ZEMeta problems fixed. CHG: Lights and materials are renamed for better writability. CHG: All light types now acts as an entity and have their own material + render command.
svn path=/branches/v0.6.1-NewGraphics/; revision=1267

+++ Source/ZEEngine/ZEGame/ZEOrbit.cpp
@@ -135,8 +135,8 @@ void ZEOrbitalMotion::Process(float ElapsedTime)
 {
 	OrbitalDay += OrbitalVelocity_AnglePerSec  * SpeedFactor * ElapsedTime;
 
-	if(OrbitalDay > ZE_PI * 2)
-		OrbitalDay = ZEMath::Mod(OrbitalDay,ZE_PI * 2);
+	if(OrbitalDay > ZE_PI * 2.0f)
+		OrbitalDay = ZEMath::Mod((float)OrbitalDay, ZE_PI * 2.0f);
 
 	AngularYear += (2 * ZE_PI / SecondsInAYear) * SpeedFactor * ElapsedTime;
